Pass port as an environment variable for healthcheck
In Windows-land we do not have container ports which we can map
externally, instead all ports are global to the machine. Healthcheck as
written in nsync uses the standard port 8080 as an argument, however
GardenWindows relies on exchanging the PORT environment variable from
the internal container as specified in requests, to the external port
before running. For this reason we would like to have the Windows
healthcheck use the PORT env variable which GardenWindows has converted,
rather than the unconverted argument.